feat(cli): add stitch-mtp command
Move stitch logic from scripts/stitch_mtp.py into
speculators.cli.stitch, wire as `speculators stitch` under the
Pipeline help panel. Original script becomes a deprecation shim.

Part of vllm-project/speculators#927 (Sub-PR 2 of 6)
