+++ tests/unit/cli/test_cli.py
@@ -15,12 +15,18 @@ class TestRootApp:
     def test_help(self):
         result = runner.invoke(app, ["--help"])
         assert result.exit_code == 0
+        assert "Pipeline" in result.output
         assert "Tools" in result.output
 
     def test_version(self):
         result = runner.invoke(app, ["--version"])
         assert result.exit_code == 0
         assert "speculators version:" in result.output
+
+    def test_pipeline_commands_in_help(self):
+        result = runner.invoke(app, ["--help"])
+        assert result.exit_code == 0
+        assert "stitch" in result.output
 
     def test_tools_commands_in_help(self):
         result = runner.invoke(app, ["--help"])
@@ -37,4 +43,16 @@ class TestConvertCommand:
 
     def test_missing_required_args(self):
         result = runner.invoke(app, ["convert"])
+        assert result.exit_code != 0
+
+
+class TestStitchCommand:
+    def test_help(self):
+        result = runner.invoke(app, ["stitch", "--help"])
+        assert result.exit_code == 0
+        assert "finetuned_checkpoint" in result.output
+        assert "verifier_path" in result.output
+
+    def test_missing_required_args(self):
+        result = runner.invoke(app, ["stitch"])
         assert result.exit_code != 0
